Moved functions from lookup service to UvidExistsValidator
