+++ src/main/java/uk/gov/companieshouse/pscverificationapi/service/impl/PscLookupServiceImpl.java
@@ -1,29 +1,19 @@
 package uk.gov.companieshouse.pscverificationapi.service.impl;
 
 import java.text.MessageFormat;
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
 
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import uk.gov.companieshouse.api.error.ApiErrorResponseException;
 import uk.gov.companieshouse.api.handler.exception.URIValidationException;
-import uk.gov.companieshouse.api.identityverification.model.UvidMatch;
 import uk.gov.companieshouse.api.model.psc.IndividualFullRecord;
 import uk.gov.companieshouse.api.model.pscverification.PscVerificationData;
 import uk.gov.companieshouse.api.model.transaction.Transaction;
-import uk.gov.companieshouse.api.psc.DateOfBirth;
-import uk.gov.companieshouse.api.psc.NameElements;
 import uk.gov.companieshouse.environment.EnvironmentReader;
 import uk.gov.companieshouse.logging.Logger;
 import uk.gov.companieshouse.pscverificationapi.enumerations.PscType;
 import uk.gov.companieshouse.pscverificationapi.exception.FilingResourceNotFoundException;
 import uk.gov.companieshouse.pscverificationapi.exception.PscLookupServiceException;
-import uk.gov.companieshouse.pscverificationapi.helper.JsonHelper;
 import uk.gov.companieshouse.pscverificationapi.sdk.companieshouse.ApiClientService;
 import uk.gov.companieshouse.pscverificationapi.service.PscLookupService;
 import uk.gov.companieshouse.pscverificationapi.utils.LogHelper;
@@ -34,7 +24,7 @@ public class PscLookupServiceImpl implements PscLookupService {
 
     private final ApiClientService apiClientService;
     private final Logger logger;
-    private EnvironmentReader environmentReader;
+    private final EnvironmentReader environmentReader;
 
     public PscLookupServiceImpl(ApiClientService apiClientService, Logger logger, EnvironmentReader environmentReader) {
         this.apiClientService = apiClientService;
@@ -93,53 +83,5 @@ public class PscLookupServiceImpl implements PscLookupService {
                     MessageFormat.format("Error Retrieving PSC details for {0}: {1}", pscAppointmentId,
                             e.getMessage()), e);
         }
-    }
-
-    /**
-     * Retrieve a UvidMatch with the PSC data.
-     *
-     * @param transaction       the Transaction
-     * @param data              the PSC verification data
-     * @param pscType           the PSC Type
-     * @return the UvidMatch, if found
-     * @throws PscLookupServiceException if the PSC was not found or an error occurred
-     */
-    @Override
-    public UvidMatch getUvidMatchWithPscData(final Transaction transaction,
-        final PscVerificationData data, final PscType pscType)
-        throws PscLookupServiceException {
-
-        UvidMatch uvidMatch = new UvidMatch();
-        Optional<String> uvid = Optional.ofNullable(data.verificationDetails().uvid());
-        uvidMatch.setUvid(uvid.orElse(""));
-
-        IndividualFullRecord pscIndividualFullRecord = getPscIndividualFullRecord(transaction, data, pscType);
-        setUvidDataFromPsc(uvidMatch, pscIndividualFullRecord);
-
-        return uvidMatch;
-    }
-
-    private void setUvidDataFromPsc(UvidMatch uvidMatch, IndividualFullRecord pscData) {
-
-        NameElements convertedNameElements = JsonHelper.convertLinkedHashmap(pscData.getNameElements(), PropertyNamingStrategies.SNAKE_CASE, NameElements.class);
-        List<String> forenames = new ArrayList<>();
-
-        String forename = convertedNameElements.getForename();
-        String middleNames = convertedNameElements.getMiddleName();
-        String surname = convertedNameElements.getSurname();
-
-        if (forename != null) {
-            forenames.add(forename);
-        }
-        if (middleNames != null) {
-            forenames.add(middleNames);
-        }
-        uvidMatch.setForenames(forenames);
-        uvidMatch.setSurname(surname);
-
-        DateOfBirth dateOfBirth = JsonHelper.convertLinkedHashmap(pscData.getDateOfBirth(), PropertyNamingStrategies.SNAKE_CASE, DateOfBirth.class);
-        String stringDateOfBirth = String.format("%04d-%02d-%02d", dateOfBirth.getYear(), dateOfBirth.getMonth(), dateOfBirth.getDay());
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-        uvidMatch.setDateOfBirth(LocalDate.parse(stringDateOfBirth, formatter));
     }
 }
