+++ src/test/java/uk/gov/companieshouse/pscverificationapi/validator/UvidExistsValidatorTest.java
@@ -15,25 +15,31 @@ import org.springframework.validation.FieldError;
 import uk.gov.companieshouse.api.error.ApiErrorResponseException;
 import uk.gov.companieshouse.api.identityverification.model.UvidMatch;
 import uk.gov.companieshouse.api.identityverification.model.UvidMatchResponse;
+import uk.gov.companieshouse.api.model.psc.IndividualFullRecord;
 import uk.gov.companieshouse.api.model.pscverification.NameMismatchReasonConstants;
 import uk.gov.companieshouse.api.model.pscverification.PscVerificationData;
 import uk.gov.companieshouse.api.model.pscverification.VerificationDetails;
 import uk.gov.companieshouse.api.model.transaction.Transaction;
+import uk.gov.companieshouse.api.psc.DateOfBirth;
+import uk.gov.companieshouse.api.psc.NameElements;
 import uk.gov.companieshouse.pscverificationapi.enumerations.PscType;
 import uk.gov.companieshouse.pscverificationapi.exception.IdvLookupServiceException;
 import uk.gov.companieshouse.pscverificationapi.service.IdvLookupService;
 import uk.gov.companieshouse.pscverificationapi.service.PscLookupService;
 
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 import java.util.*;
 import java.util.stream.Stream;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.*;
 import static org.hamcrest.collection.IsEmptyCollection.empty;
 import static org.hamcrest.collection.IsIterableContainingInOrder.contains;
 import static org.mockito.Mockito.when;
 import static uk.gov.companieshouse.api.identityverification.model.UvidMatchResponse.AccuracyStatementEnum.*;
+import static uk.gov.companieshouse.pscverificationapi.enumerations.PscType.INDIVIDUAL;
 
 @ExtendWith(MockitoExtension.class)
 class UvidExistsValidatorTest {
@@ -56,23 +62,36 @@ class UvidExistsValidatorTest {
     @Mock
     private UvidMatchResponse uvidMatchResponse;
 
-    @Mock
-    private VerificationDetails verificationDetails;
-
-    @Mock
-    private UvidMatch uvidMatch;
-
     UvidExistsValidator testValidator;
     private PscType pscType;
     private Set<FieldError> errors;
     private String passthroughHeader;
     VerificationValidationContext validationContext;
 
+    private static final String PSC_ID = "654321";
+    private static final String UVID_CODE = "XY222222223";
+    private static final String COMPANY_NUMBER = "12345678";
+    private static final VerificationDetails VERIFICATION_DETAILS =
+        VerificationDetails.newBuilder().uvid(UVID_CODE)
+            .build();
+    private static final PscVerificationData PSC_VERIFICATION_DATA =
+        PscVerificationData.newBuilder()
+            .pscAppointmentId(PSC_ID)
+            .companyNumber(COMPANY_NUMBER)
+            .verificationDetails(VERIFICATION_DETAILS)
+            .build();
+    private static final String[] FORENAMES = {"Forename1", "Forename2 Forename3"};
+    private static final String SURNAME = "Surname";
+    private static final String PSC_FORENAME = "Forename1";
+    private static final String PSC_MIDDLE_NAME = "Forename2 Forename3";
+    private static final String PSC_SURNAME = "Surname";
+    private static final DateOfBirth DATE_OF_BIRTH = new DateOfBirth(2, 1983).day(27);
+
     @BeforeEach
     void setUp() {
 
         errors = new HashSet<>();
-        pscType = PscType.INDIVIDUAL;
+        pscType = INDIVIDUAL;
         passthroughHeader = "passthroughHeader";
         validationContext = new VerificationValidationContext(pscVerificationData, errors, transaction, pscType, passthroughHeader);
 
@@ -81,59 +100,115 @@ class UvidExistsValidatorTest {
     }
 
     @Test
-    void validateWhenUvidMatch() throws ApiErrorResponseException {
-
-        when(pscLookupService.getUvidMatchWithPscData(transaction, pscVerificationData, pscType))
-            .thenReturn(uvidMatch);
-        when(idvLookupService.matchUvid(uvidMatch)).thenReturn(uvidMatchResponse);
-        when(uvidMatchResponse.getAccuracyStatement()).thenReturn(Collections.singletonList(DETAILS_MATCH_UVID));
-
-        testValidator.validate(
-            new VerificationValidationContext(pscVerificationData, errors, transaction, pscType, passthroughHeader)
-        );
-        assertThat(errors, is(empty()));
-    }
-
-    @Test
     void validateWhenException() throws ApiErrorResponseException {
 
-        when(pscVerificationData.verificationDetails()).thenReturn(verificationDetails);
-        when(pscVerificationData.verificationDetails().uvid()).thenReturn("1234567890");
-        when(pscLookupService.getUvidMatchWithPscData(transaction, pscVerificationData, pscType))
-            .thenReturn(uvidMatch);
-        when(idvLookupService.matchUvid(uvidMatch)).thenThrow(new ApiErrorResponseException(
+        UvidMatch expected = createUvid(UVID_CODE);
+        setNames(expected, FORENAMES, SURNAME);
+
+        String stringDateOfBirth = String.format("%04d-%02d-%02d", 1983, 2, 27);
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+        expected.setDateOfBirth(LocalDate.parse(stringDateOfBirth, formatter));
+
+        final var pscData = createPscData(PSC_FORENAME, PSC_MIDDLE_NAME, PSC_SURNAME, DATE_OF_BIRTH);
+
+        when(pscLookupService.getPscIndividualFullRecord(transaction, PSC_VERIFICATION_DATA, pscType)).thenReturn(pscData);
+        when(idvLookupService.matchUvid(expected)).thenThrow(new ApiErrorResponseException(
                 new HttpResponseException.Builder(400, "test error", new HttpHeaders())));
 
         final var exception = Assertions.assertThrows(IdvLookupServiceException.class,
-            () -> testValidator.validate(new VerificationValidationContext(pscVerificationData, errors, transaction, pscType, passthroughHeader)));
-        assertThat(exception.getMessage(), is("Error matching UVID 1234567890: null test error"));
+            () -> testValidator.validate(new VerificationValidationContext(PSC_VERIFICATION_DATA, errors, transaction, pscType, passthroughHeader)));
+        assertThat(exception.getMessage(), is("Error matching UVID XY222222223: null test error"));
     }
 
     @Test
     void validateWhenNameMismatchNotNull() throws ApiErrorResponseException {
 
+        UvidMatch expected = createUvid(UVID_CODE);
+        setNames(expected, FORENAMES, SURNAME);
+
+        String stringDateOfBirth = String.format("%04d-%02d-%02d", 1983, 2, 27);
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+        expected.setDateOfBirth(LocalDate.parse(stringDateOfBirth, formatter));
+
+        var verificationDetails = VerificationDetails.newBuilder(PSC_VERIFICATION_DATA.verificationDetails()).nameMismatchReason(NameMismatchReasonConstants.PREFERRED_NAME).build();
+
+        final var individualFullRecord = createPscData(PSC_FORENAME, PSC_MIDDLE_NAME, PSC_SURNAME, DATE_OF_BIRTH);
+
+        when(pscLookupService.getPscIndividualFullRecord(transaction, pscVerificationData, pscType)).thenReturn(individualFullRecord);
         when(pscVerificationData.verificationDetails()).thenReturn(verificationDetails);
-        when(pscVerificationData.verificationDetails().nameMismatchReason()).thenReturn(NameMismatchReasonConstants.PREFERRED_NAME);
-        when(pscLookupService.getUvidMatchWithPscData(transaction, pscVerificationData, pscType))
-            .thenReturn(uvidMatch);
-        when(idvLookupService.matchUvid(uvidMatch)).thenReturn(uvidMatchResponse);
+        when(idvLookupService.matchUvid(expected)).thenReturn(uvidMatchResponse);
         when(uvidMatchResponse.getAccuracyStatement()).thenReturn(Collections.singletonList(FORENAMES_MISMATCH));
 
         testValidator.validate(validationContext);
+    }
+
+    @Test
+    void validateWhenUvidMatch() throws ApiErrorResponseException {
+
+        UvidMatch expected = createUvid(UVID_CODE);
+        setNames(expected, FORENAMES, SURNAME);
+
+        String stringDateOfBirth = String.format("%04d-%02d-%02d", 1983, 2, 27);
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+        expected.setDateOfBirth(LocalDate.parse(stringDateOfBirth, formatter));
+
+        final var individualFullRecord = createPscData(PSC_FORENAME, PSC_MIDDLE_NAME, PSC_SURNAME, DATE_OF_BIRTH);
+        when(pscLookupService.getPscIndividualFullRecord(transaction, pscVerificationData, pscType)).thenReturn(individualFullRecord);
+
+        when(pscVerificationData.verificationDetails()).thenReturn(VERIFICATION_DETAILS);
+        when(idvLookupService.matchUvid(expected)).thenReturn(uvidMatchResponse);
+        when(uvidMatchResponse.getAccuracyStatement()).thenReturn(Collections.singletonList(DETAILS_MATCH_UVID));
+
+        testValidator.validate(validationContext);
+    }
+
+    @Test
+    void getUvidMatchWithPscDataNoForename() throws ApiErrorResponseException {
+
+        UvidMatch expected = createUvid(UVID_CODE);
+        setNames(expected, null, SURNAME);
+
+        String stringDateOfBirth = String.format("%04d-%02d-%02d", 1983, 2, 27);
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+        expected.setDateOfBirth(LocalDate.parse(stringDateOfBirth, formatter));
+
+        final var pscData = createPscData(null, null, PSC_SURNAME, DATE_OF_BIRTH);
+
+        when(pscLookupService.getPscIndividualFullRecord(transaction, PSC_VERIFICATION_DATA, pscType)).thenReturn(pscData);
+
+        var testUvid = testValidator.getUvidMatchWithPscData(transaction, PSC_VERIFICATION_DATA, INDIVIDUAL);
+
+        assertThat("Forenames order does not match", testUvid.getForenames(), is(equalTo(expected.getForenames())));
+        assertThat(testUvid, samePropertyValuesAs(expected));
+
+        when(idvLookupService.matchUvid(expected)).thenReturn(uvidMatchResponse);
+        when(uvidMatchResponse.getAccuracyStatement()).thenReturn(Collections.singletonList(DETAILS_MATCH_UVID));
+
+        testValidator.validate(
+            new VerificationValidationContext(PSC_VERIFICATION_DATA, errors, transaction, pscType, passthroughHeader)
+        );
+        assertThat(errors, is(empty()));
     }
 
     @ParameterizedTest
     @MethodSource("provideParameters")
     void validateWhenUvidMatchError(String errorValue, String errorText, List<UvidMatchResponse.AccuracyStatementEnum> statementEnums) throws ApiErrorResponseException {
 
-        when(pscVerificationData.verificationDetails()).thenReturn(verificationDetails);
-        when(pscVerificationData.verificationDetails().uvid()).thenReturn(errorValue);
-        when(pscLookupService.getUvidMatchWithPscData(transaction, pscVerificationData, pscType))
-            .thenReturn(uvidMatch);
-        when(idvLookupService.matchUvid(uvidMatch)).thenReturn(uvidMatchResponse);
+        UvidMatch expected = createUvid(UVID_CODE);
+        setNames(expected, FORENAMES, SURNAME);
+
+        String stringDateOfBirth = String.format("%04d-%02d-%02d", 1983, 2, 27);
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+        expected.setDateOfBirth(LocalDate.parse(stringDateOfBirth, formatter));
+
+        final var individualFullRecord = createPscData(PSC_FORENAME, PSC_MIDDLE_NAME, PSC_SURNAME, DATE_OF_BIRTH);
+        when(pscLookupService.getPscIndividualFullRecord(transaction, pscVerificationData, pscType)).thenReturn(individualFullRecord);
+
+        when(pscVerificationData.verificationDetails()).thenReturn(VERIFICATION_DETAILS);
+        when(idvLookupService.matchUvid(expected)).thenReturn(uvidMatchResponse);
         when(uvidMatchResponse.getAccuracyStatement()).thenReturn(statementEnums);
 
-        var fieldError = new FieldError("object", "uvid_match", pscVerificationData.verificationDetails().uvid(), false,
+        var fieldError = new FieldError("object", "uvid_match", PSC_VERIFICATION_DATA.verificationDetails().uvid(), false,
             new String[]{null, errorValue}, null, errorText);
 
         when(validation.get(errorValue)).thenReturn(errorText);
@@ -160,5 +235,34 @@ class UvidExistsValidatorTest {
             Arguments.of("no-name-mismatch-reason",
                 "The name on the public register is different to the name this PSC used for identity verification:" +
                     " a name mismatch reason must be provided", List.of(FORENAMES_MISMATCH, SURNAME_MISMATCH)));
+    }
+
+    private static UvidMatch createUvid(String uvidCode) {
+        UvidMatch uvidMatch = new UvidMatch();
+        uvidMatch.setUvid(uvidCode);
+        return uvidMatch;
+    }
+
+    private static void setNames(final UvidMatch uvidMatch, String[] forenames, String surname) {
+        List<String> forenamesAsList = Optional.ofNullable(forenames)
+            .map(Arrays::asList)
+            .map(ArrayList::new)
+            .orElseGet(ArrayList::new);
+
+        uvidMatch.setForenames(forenamesAsList);
+        uvidMatch.setSurname(surname);
+    }
+
+    private static IndividualFullRecord createPscData(String forename, String middleName, String surname, DateOfBirth dateOfBirth) {
+        IndividualFullRecord individualFullRecord = new IndividualFullRecord();
+        NameElements nameElements = new NameElements();
+        nameElements.setForename(forename);
+        nameElements.setMiddleName(middleName);
+        nameElements.setSurname(surname);
+        individualFullRecord.setNameElements(nameElements);
+
+        individualFullRecord.setDateOfBirth(dateOfBirth);
+
+        return individualFullRecord;
     }
 }
